Fin Table Depense

+++ PlayZone.BLL/Models/Budget_Related/Depense.cs
@@ -6,7 +6,7 @@ public class Depense
     public int LibeleId { get; set; }
     public int ProjectId { get; set; }
     public int OrganismeId {get; set;}
-    public int Montant { get; set; }
+    public Decimal Montant { get; set; }
     public DateTime DateIntervention { get; set; }
     public DateTime DateFacturation { get; set; }
     public string Motif { get; set; }
